refactor: 추천 기능 recommend 파일 리팩토링
