+++ src/main/java/menu/MessageView.java
@@ -12,7 +12,7 @@ public class MessageView {
     }
 
     public void recommendMessage(){
-        System.out.println("[ 구분 | 월요일 | 화요일 | 수요일 | 목요일 | 금요일 ]\n");
-        System.out.println("[ 카테고리 | 한식 | 한식 | 일식 | 중식 | 아시안 ]\n");
+        System.out.println("[ 구분 | 월요일 | 화요일 | 수요일 | 목요일 | 금요일 ]");
+        System.out.println("[ 카테고리 | 한식 | 한식 | 일식 | 중식 | 아시안 ]");
     }
 }
